ticket/requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.23
gunicorn==20.0.4